# core/vm: rework jumpdest analysis benchmarks