--- a/core/vm/analysis_test.go
+++ b/core/vm/analysis_test.go
@@ -89,6 +89,9 @@ func BenchmarkJumpdestOpAnalysis(bench *testing.B) {
 		bits := make(bitvec, len(code)/8+1+4)
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
+			for j := range bits {
+				bits[j] = 0
+			}
 			codeBitmapInternal(code, bits)
 		}
 	}