@chfast 2021-08-30
Comparison without and with bitvec clearing in the benchmark loop:
```
name                           old time/op    new time/op    delta                                                                                                                                                                
JumpdestOpAnalysis/PUSH1-8        979µs ± 0%     983µs ± 0%  +0.42%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH2-8        943µs ± 0%     947µs ± 0%  +0.40%  (p=0.000 n=8+8)                                                                                                                                              
JumpdestOpAnalysis/PUSH3-8        694µs ± 0%     698µs ± 0%  +0.55%  (p=0.000 n=10+10)                                                                                                                                            
JumpdestOpAnalysis/PUSH4-8        581µs ± 0%     584µs ± 0%  +0.62%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH5-8        489µs ± 0%     493µs ± 0%  +0.76%  (p=0.000 n=8+8)                                                                                                                                              
JumpdestOpAnalysis/PUSH6-8        464µs ± 0%     468µs ± 0%  +0.80%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH7-8        350µs ± 0%     353µs ± 0%  +1.04%  (p=0.000 n=9+8)                                                                                                                                              
JumpdestOpAnalysis/PUSH8-8        405µs ± 0%     409µs ± 0%  +0.88%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH9-8        362µs ± 0%     366µs ± 0%  +1.11%  (p=0.000 n=8+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH10-8       434µs ± 0%     438µs ± 0%  +0.93%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH11-8       419µs ± 0%     423µs ± 0%  +0.90%  (p=0.000 n=9+8)                                                                                                                                              
JumpdestOpAnalysis/PUSH12-8       352µs ± 0%     356µs ± 0%  +1.05%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH13-8       330µs ± 0%     341µs ± 5%  +3.58%  (p=0.000 n=9+9)                                                                                                                                              
JumpdestOpAnalysis/PUSH14-8       338µs ± 0%     342µs ± 0%  +1.10%  (p=0.000 n=9+9)
JumpdestOpAnalysis/PUSH15-8       332µs ± 1%     336µs ± 0%  +1.03%  (p=0.000 n=9+10)
JumpdestOpAnalysis/PUSH16-8       241µs ± 4%     247µs ± 2%  +2.80%  (p=0.009 n=10+10)
JumpdestOpAnalysis/PUSH17-8       211µs ± 0%     221µs ± 0%  +4.93%  (p=0.000 n=8+9)
JumpdestOpAnalysis/PUSH18-8       266µs ± 0%     270µs ± 0%  +1.36%  (p=0.000 n=9+8)
JumpdestOpAnalysis/PUSH19-8       266µs ± 0%     269µs ± 0%  +1.41%  (p=0.000 n=8+9)
JumpdestOpAnalysis/PUSH20-8       231µs ± 0%     235µs ± 0%  +1.59%  (p=0.000 n=9+9)
JumpdestOpAnalysis/PUSH21-8       222µs ± 0%     226µs ± 0%  +1.67%  (p=0.000 n=8+8)
JumpdestOpAnalysis/PUSH22-8       233µs ± 0%     237µs ± 0%  +1.65%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH23-8       233µs ± 0%     237µs ± 0%  +1.58%  (p=0.000 n=9+9)
JumpdestOpAnalysis/PUSH24-8       213µs ± 0%     216µs ± 0%  +1.70%  (p=0.000 n=9+9)
JumpdestOpAnalysis/PUSH25-8       215µs ± 0%     219µs ± 0%  +1.73%  (p=0.000 n=9+8)
JumpdestOpAnalysis/PUSH26-8       239µs ± 0%     243µs ± 0%  +1.55%  (p=0.000 n=9+8)
JumpdestOpAnalysis/PUSH27-8       240µs ± 0%     243µs ± 0%  +1.54%  (p=0.000 n=9+8)
JumpdestOpAnalysis/PUSH28-8       216µs ± 0%     220µs ± 0%  +1.66%  (p=0.000 n=7+10)
JumpdestOpAnalysis/PUSH29-8       210µs ± 0%     213µs ± 0%  +1.69%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH30-8       218µs ± 0%     221µs ± 0%  +1.70%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH31-8       218µs ± 0%     222µs ± 0%  +1.72%  (p=0.000 n=9+9)
JumpdestOpAnalysis/PUSH32-8       169µs ± 0%     173µs ± 0%  +2.16%  (p=0.000 n=9+8)
JumpdestOpAnalysis/JUMPDEST-8     490µs ± 0%     493µs ± 0%  +0.69%  (p=0.000 n=10+9)
JumpdestOpAnalysis/STOP-8         490µs ± 0%     493µs ± 0%  +0.61%  (p=0.000 n=10+8)
```